cleanup airtable plugin
